reorganise unit tests dir
